Move elastic profiles /usages api to internal path
