+++ api/api-elastic-profile-v1/src/main/java/com/thoughtworks/go/apiv1/elasticprofile/ElasticProfileControllerV1.java
@@ -23,12 +23,10 @@ import com.thoughtworks.go.api.base.OutputWriter;
 import com.thoughtworks.go.api.spring.ApiAuthenticationHelper;
 import com.thoughtworks.go.api.util.GsonTransformer;
 import com.thoughtworks.go.apiv1.elasticprofile.representers.ElasticProfileRepresenter;
-import com.thoughtworks.go.apiv1.elasticprofile.representers.ElasticProfileUsageRepresenter;
 import com.thoughtworks.go.apiv1.elasticprofile.representers.ElasticProfilesRepresenter;
 import com.thoughtworks.go.config.PluginProfiles;
 import com.thoughtworks.go.config.elastic.ElasticProfile;
 import com.thoughtworks.go.config.exceptions.RecordNotFoundException;
-import com.thoughtworks.go.domain.ElasticProfileUsage;
 import com.thoughtworks.go.server.service.ElasticProfileService;
 import com.thoughtworks.go.server.service.EntityHashingService;
 import com.thoughtworks.go.server.service.result.HttpLocalizedOperationResult;
@@ -41,7 +39,6 @@ import spark.Request;
 import spark.Response;
 
 import java.io.IOException;
-import java.util.Collection;
 import java.util.function.Consumer;
 
 import static com.thoughtworks.go.api.util.HaltApiResponses.*;
@@ -78,7 +75,7 @@ public class ElasticProfileControllerV1 extends ApiController implements SparkSp
 
             get("", mimeType, this::index);
             get(Routes.ElasticProfileAPI.ID, mimeType, this::show);
-            get(Routes.ElasticProfileAPI.ID + Routes.ElasticProfileAPI.USAGES, mimeType, this::usages);
+
             post("", mimeType, this::create);
             put(Routes.ElasticProfileAPI.ID, mimeType, this::update);
             delete(Routes.ElasticProfileAPI.ID, mimeType, this::destroy);
@@ -141,11 +138,6 @@ public class ElasticProfileControllerV1 extends ApiController implements SparkSp
         return renderHTTPOperationResult(result, request, response);
     }
 
-    public String usages(Request request, Response response) {
-        final String elasticProfileId = StringUtils.stripToEmpty(request.params(PROFILE_ID_PARAM));
-        final Collection<ElasticProfileUsage> jobsUsingElasticProfile = elasticProfileService.getUsageInformation(elasticProfileId);
-        return ElasticProfileUsageRepresenter.toJSON(jobsUsingElasticProfile);
-    }
 
 
     private boolean isRenameAttempt(String profileIdFromRequestParam, String profileIdFromRequestBody) {
