+++ server/webapp/WEB-INF/applicationContext-global.xml
@@ -27,6 +27,7 @@
 
   <context:annotation-config/>
 
+  <context:component-scan base-package="com.thoughtworks.go.apiv1.elasticprofileoperation"/>
   <context:component-scan base-package="com.thoughtworks.go.apiv1.usersearch"/>
   <context:component-scan base-package="com.thoughtworks.go.apiv1.serverdrainmode"/>
   <context:component-scan base-package="com.thoughtworks.go.apiv4.agents"/>
