+++ settings.gradle
@@ -36,6 +36,7 @@ include ':api:api-dashboard-v2'
 include ':api:api-data-sharing-reporting-v2'
 include ':api:api-data-sharing-settings-v1'
 include ':api:api-data-sharing-usage-data-v3'
+include ':api:api-elastic-profile-operation-v1'
 include ':api:api-elastic-profile-v1'
 include ':api:api-encryption-v1'
 include ':api:api-material-search-v1'
